# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gachita API",
        "title": "Gachita API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "내 알림 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/NotificationResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/read-all": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "알림 전체 읽음 처리",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/notifications/{notificationId}/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "알림 읽음 처리",
                "parameters": [
                    {
                        "type": "string",
                        "description": "알림 ID",
                        "name": "notificationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/NotificationResponse"
                        }
                    }
                }
            }
        },
...
        "/api/rooms/{id}/queue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "매칭 대기 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/QueueEntryResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "매칭 대기 등록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "대기 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateQueueEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/QueueEntryResponse"
                        }
                    }
                }
            }
        },
        "/api/rooms/{id}/queue/{entryId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "매칭 대기 취소",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "대기 ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/QueueEntryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateQueueEntryRequest": {
            "type": "object",
            "required": [
                "from_stop_id",
                "time_end",
                "time_start",
                "to_stop_id"
            ],
            "properties": {
                "from_stop_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "max_seats": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 4
                },
                "min_seats": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 2
                },
                "time_end": {
                    "type": "string",
                    "example": "2026-08-02T10:00:00Z"
                },
                "time_start": {
                    "type": "string",
                    "example": "2026-08-02T09:00:00Z"
                },
                "to_stop_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                }
            }
        },
...
        "MessageResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "NotificationResponse": {
            "type": "object",
            "required": [
                "body",
                "created_at",
                "id",
                "title",
                "type"
            ],
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "read_at": {
                    "type": "string",
                    "x-nullable": true
                },
                "ref_id": {
                    "type": "string",
                    "x-nullable": true
                },
                "ref_type": {
                    "type": "string",
                    "x-nullable": true
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "QueueEntryResponse": {
            "type": "object",
            "required": [
                "from_stop_id",
                "id",
                "max_seats",
                "min_seats",
                "room_id",
                "status",
                "time_end",
                "time_start",
                "to_stop_id",
                "user_id"
            ],
            "properties": {
                "from_stop_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string",
                    "x-nullable": true
                },
                "max_seats": {
                    "type": "integer",
                    "example": 4
                },
                "min_seats": {
                    "type": "integer",
                    "example": 2
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "waiting",
                        "matched",
                        "cancelled"
                    ]
                },
                "time_end": {
                    "type": "string",
                    "example": "2026-08-02T10:00:00Z"
                },
                "time_start": {
                    "type": "string",
                    "example": "2026-08-02T09:00:00Z"
                },
                "to_stop_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}